NEOABZU/pyproject.toml
[build-system]
requires = ["maturin>=1.5,<2"]
build-backend = "maturin"

[project]
name = "neoabzu"
version = "0.1.0"
description = "Python helpers for the NEOABZU Rust workspace"
requires-python = ">=3.10"
authors = [{name = "NEOABZU Team"}]
dependencies = ["grpcio>=1.62"]

[project.urls]
"Blueprint Spine" = "../docs/blueprint_spine.md"
"The Absolute Protocol" = "../docs/The_Absolute_Protocol.md"

[tool.maturin]
python-packages = ["neoabzu"]

[[tool.maturin.targets]]
name = "neoabzu_core"
path = "core"

[[tool.maturin.targets]]
name = "neoabzu_memory"
path = "memory"

[[tool.maturin.targets]]
name = "neoabzu_vector"
path = "vector"

[[tool.maturin.targets]]
name = "neoabzu_numeric"
path = "numeric"

[[tool.maturin.targets]]
name = "neoabzu_fusion"
path = "fusion"

[[tool.maturin.targets]]
name = "neoabzu_persona"
path = "persona"

[[tool.maturin.targets]]
name = "neoabzu_crown"
path = "crown"

[[tool.maturin.targets]]
name = "neoabzu_rag"
path = "rag"

[[tool.maturin.targets]]
name = "neoabzu_insight"
path = "insight"

[tool.neoabzu.workspace]
members = ["core", "memory", "vector", "fusion", "persona", "crown", "rag", "numeric", "insight"]
